backend/pyproject.toml
[project]
name = "faux-backend"
version = "0.1.0"
description = "Faux backend - Smart field classification and data generation"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "faker>=33.0.0",
    "google-genai>=1.0.0",
    "openai>=1.40.0",
    "pgvector>=0.3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "rapidfuzz>=3.10.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
